Pass in source directories and not source files to ExtractRangeMap tasks
SourceDirectorySet extends FileTree, which enumerates all the source files in the
source directories, causing issues with Srg2Source expecting directories and not individual
files.

Fixed by passing in SourceDirectorySet#getSourceDirectories, which preserves the lazy
nature (the return FileCollection is a live view, as opposed to #getSrcDirs which
returns a static Set<File>).